update callers, fix tests
