+++ src/test/spec/unified_runner/operation.rs
@@ -2346,7 +2346,7 @@ impl TestOperation for CreateChangeStream {
             let target = entities.get(id).unwrap();
             let stream = match target {
                 Entity::Client(ce) => {
-                    ce.watch(self.pipeline.clone(), self.options.clone())
+                    ce.watch(self.pipeline.clone()).with_options(self.options.clone())
                         .await?
                 }
                 Entity::Database(db) => {
